chore: move reset to SignForm
